+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerLicenseServiceImplTest.java
@@ -1,0 +1,79 @@
+package de.codescape.jira.plugins.scrumpoker.service;
+
+import com.atlassian.upm.api.license.PluginLicenseManager;
+import com.atlassian.upm.api.license.entity.LicenseError;
+import com.atlassian.upm.api.license.entity.PluginLicense;
+import com.atlassian.upm.api.util.Option;
+import org.junit.Rule;
+import org.junit.Test;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnit;
+import org.mockito.junit.MockitoRule;
+
+import static de.codescape.jira.plugins.scrumpoker.service.ScrumPokerLicenseServiceImpl.MISSING_LICENSE;
+import static org.hamcrest.Matchers.*;
+import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+public class ScrumPokerLicenseServiceImplTest {
+
+    @Rule
+    public MockitoRule rule = MockitoJUnit.rule();
+
+    @Mock
+    private PluginLicenseManager pluginLicenseManager;
+
+    @InjectMocks
+    private ScrumPokerLicenseServiceImpl scrumPokerLicenseService;
+
+    /* tests for hasValidLicense() */
+
+    @Test
+    public void hasValidLicenseReturnsFalseForMissingLicense() {
+        when(pluginLicenseManager.getLicense()).thenReturn(Option.none());
+        assertThat(scrumPokerLicenseService.hasValidLicense(), is(false));
+    }
+
+    @Test
+    public void hasValidLicenseReturnsFalseForLicenseWithError() {
+        PluginLicense pluginLicense = mock(PluginLicense.class);
+        when(pluginLicense.getError()).thenReturn(Option.option(LicenseError.EXPIRED));
+        when(pluginLicenseManager.getLicense()).thenReturn(Option.option(pluginLicense));
+        assertThat(scrumPokerLicenseService.hasValidLicense(), is(false));
+    }
+
+    @Test
+    public void hasValidLicenseReturnsTrueForLicenseWithoutErrors() {
+        PluginLicense pluginLicense = mock(PluginLicense.class);
+        when(pluginLicense.getError()).thenReturn(Option.none());
+        when(pluginLicenseManager.getLicense()).thenReturn(Option.option(pluginLicense));
+        assertThat(scrumPokerLicenseService.hasValidLicense(), is(true));
+    }
+
+    /* tests for getLicenseError() */
+
+    @Test
+    public void getLicenseErrorReturnsMissingForMissingLicense() {
+        when(pluginLicenseManager.getLicense()).thenReturn(Option.none());
+        assertThat(scrumPokerLicenseService.getLicenseError(), is(equalTo(MISSING_LICENSE)));
+    }
+
+    @Test
+    public void getLicenseErrorReturnsErrorForLicenseWithError() {
+        PluginLicense pluginLicense = mock(PluginLicense.class);
+        when(pluginLicense.getError()).thenReturn(Option.option(LicenseError.EXPIRED));
+        when(pluginLicenseManager.getLicense()).thenReturn(Option.option(pluginLicense));
+        assertThat(scrumPokerLicenseService.getLicenseError(), is(equalTo(LicenseError.EXPIRED.name())));
+    }
+
+    @Test
+    public void getLicenseErrorReturnsNullForLicenseWithoutErrors() {
+        PluginLicense pluginLicense = mock(PluginLicense.class);
+        when(pluginLicense.getError()).thenReturn(Option.none());
+        when(pluginLicenseManager.getLicense()).thenReturn(Option.option(pluginLicense));
+        assertThat(scrumPokerLicenseService.getLicenseError(), is(nullValue()));
+    }
+
+}