+++ docs/changelog.md
@@ -11,7 +11,7 @@ category: Administration
 
 ### [Unreleased]
 
-* ...
+* refactor: extract license check into separate reusable service class
 
 ### [20.10.1] - 2020-10-29